chore: moving server logic + cleanup imports
