+++ core/threshold/Cargo.toml
@@ -10,7 +10,6 @@ license.workspace = true
 # ⚠️ IMPORTANT: Add new dependencies to workspace root Cargo.toml, NOT here!
 # Use .workspace = true to reference workspace dependencies
 # Maintain alphabetical order
-aes.workspace = true
 aes-prng.workspace = true
 algebra = { workspace = true, default-features = false }
 anyhow.workspace = true
@@ -19,7 +18,7 @@ bc2wrap.workspace = true
 clap = { workspace = true, features = ["derive", "env"] }
 dashmap.workspace = true
 error-utils.workspace = true
-threshold-execution = { workspace = true, features = ["malicious_strategies"]}
+threshold-execution = { workspace = true, features = ["malicious_strategies"] }
 futures.workspace = true
 futures-util.workspace = true
 getrandom.workspace = true
@@ -29,14 +28,12 @@ k256.workspace = true
 minijinja = { workspace = true, optional = true }
 observability.workspace = true
 threshold-networking.workspace = true
-paste.workspace = true
 peak_alloc = { workspace = true, optional = true }
 rand.workspace = true
 serde = { workspace = true, features = ["derive", "rc"] }
 sha2.workspace = true
 strum = { workspace = true, features = ["derive"] }
 strum_macros.workspace = true
-test-utils = { workspace = true, optional = true }
 tfhe = { workspace = true, features = [
     "boolean",
     "shortint",
@@ -50,29 +47,27 @@ tokio = { workspace = true, features = [
     "rt",
     "macros",
     "time",
-]}
-tokio-rustls = { workspace = true }
+] }
+tokio-rustls.workspace = true
 tonic = { workspace = true, features = ["tls-ring"] }
-tonic-health = { workspace = true }
+tonic-health.workspace = true
 tower = { workspace = true, features = ["retry", "timeout"] }
 tower-http = { workspace = true, features = ["trace"] }
-threshold-experimental = { workspace = true }
+threshold-experimental.workspace = true
 threshold-types.workspace = true
 tracing = { workspace = true, features = ["log"] }
-x509-parser = { workspace = true }
+x509-parser.workspace = true
 zeroize.workspace = true
-
-[build-dependencies]
-tonic-prost-build.workspace = true
 
 [dev-dependencies]
 # ⚠️ IMPORTANT: Add new dependencies to workspace root Cargo.toml, NOT here!
 # Use .workspace = true to reference workspace dependencies
 # Maintain alphabetical order
-criterion= {workspace = true, features = ["async_tokio", "html_reports"] }
+criterion = { workspace = true, features = ["async_tokio", "html_reports"] }
 crypto-bigint.workspace = true
 threshold-execution = { workspace = true, features = ["testing"] }
 threshold-experimental.workspace = true
+paste.workspace = true
 pprof.workspace = true
 rayon.workspace = true
 redis.workspace = true
@@ -240,7 +235,7 @@ harness = false
 [features]
 default = ["extension_degree_4"]
 templating = ["minijinja"]
-measure_memory = ["peak_alloc"]
+measure_memory = ["dep:peak_alloc"]
 
 extension_degree_3 = ["algebra/extension_degree_3"]
 extension_degree_4 = ["algebra/extension_degree_4"]
