+++ core/threshold/src/choreography/mod.rs
@@ -1,4 +1,5 @@
 pub mod client_utils;
+pub mod server;
 pub mod server_utils;
 pub mod tfhe_rs;
 pub mod bgv;
