+++ core/threshold/src/lib.rs
@@ -2,6 +2,5 @@
 pub mod allocator;
 pub mod choreography;
 pub mod conf;
-pub mod grpc;
 pub mod utils;
 pub mod zk_utils;
